requirements.txt
streamlit
python-dotenv
langchain
faiss-cpu
sentence-transformers
numpy
huggingface_hub
geopy
scipy
serpapi